Phrasing changes
By line number:
3: This utility doesn't run the games, so clarifying that it
merely streamlines patch installation is important.

9: Steam Cloud saves are a feature, and they cannot
own their own implementation. Thus, use of an apostrophe
is incorrect. The intent of that change is clear, though, so
the phrase now reads "implementation of".
Also, removing "for the title in question" is problematic,
since the implementation of Steam Cloud saves itself is not the
problematic part in most cases. Managers of a Steam AppID have to
implement Steam Cloud saves properly in the game and
its configuration. Lack of proper implementation for a given
game is the most common issue with this feature.
The removal of that phrase has been undone.

19: Removing the repetitive use of the word "either" is a good
choice. In actuality, both instances were unnecessary
and reduced clarity. The second instance has been removed.

65: The request for PRs was indeed redundant. However,
it makes sense to instead specify the distros that were
tested against and indicate that PRs for other distros
are "especially appreciated". PRs meaning pull requests may
not be immediately obvious to users, so "pull request" is now
used in place of "PR".